src/stat-service/requirements.txt
flask>=3.0.3
clickhouse-connect>=0.7.12
kafka-python>=2.0.2
